gate + partie blog et articles

+++ app/Http/Controllers/ImageController.php
@@ -16,7 +16,7 @@ class ImageController extends Controller
      */
     public function index()
     {
-        // $this->authorize("admin");
+        $this->authorize("admin");
         $images = Image::all();
         return view('admin.image.main', compact('images'));
     }
@@ -40,7 +40,7 @@ class ImageController extends Controller
      */
     public function store(Request $request)
     {
-        // $this->authorize("admin");
+        $this->authorize("admin");
         request()->validate([
             "nom" => ["required"],
             // "src" => ['required', 'image', 'mimes:jpeg,png,jpg,gif,svg', 'max:2048'],    
@@ -107,7 +107,7 @@ class ImageController extends Controller
      */
     public function destroy(Image $image)
     {
-        // $this->authorize("admin");
+        $this->authorize("admin");
         // Storage::disk('public')->delete('img/' . $image->src);
         $image->delete();
         return redirect()->route('image.index')->with('warning', 'Image bien supprimé');
